% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManFunks.R
\name{addAlpha}
\alias{addAlpha}
\title{Add alpha channel (transparency) to a color.}
\usage{
addAlpha(col, alpha = 0.25)
}
\arguments{
\item{col}{The color to add transparency to}

\item{alpha}{A value between zero and one, the higher the more transparency.
Default is \code{.25}.}
}
\value{
The modified color \code{col}.
}
\description{
Add alpha channel (transparency) to a color.
}
