% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManFunks.R
\name{addAncestors}
\alias{addAncestors}
\title{Adds all GO Terms that are ancestral to the argument \code{go.terms}.}
\usage{
addAncestors(go.terms, go.obo = getOption("MapMan2GO.go.obo", GO.OBO))
}
\arguments{
\item{go.terms}{a character vector of GO identifier, e.g.
\code{'GO:006969'}.}

\item{go.obo}{The result of reading the Gene Ontology OBO file with
\code{ontologyIndex::get_ontology('go.obo')}. Default is
\code{getOption('MapMan2GO.go.obo', GO.OBO)}. See \code{MapMan2GO} data for
more details.}
}
\value{
A character vector including the argument \code{go.terms} and all
found ancestors. Returns \code{character(0)} if the argument \code{go.terms}
is NULL, all of its entries are \code{NA}, or if its length equals zero.
}
\description{
Adds all GO Terms that are ancestral to the argument \code{go.terms}.
}
