% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManFunks.R
\name{analyzeSharedWords}
\alias{analyzeSharedWords}
\title{Analyze the number of shared words among the MapMan-Bins descriptions and the
GOs that conform their GOA.}
\usage{
analyzeSharedWords(mm.2.go.df.MapManBin.GO, mm.2.go.df.MapManBin)
}
\arguments{
\item{mm.2.go.df.MapManBin.GO}{a list of the GOAs to analyze that conform the
MapMan-Bin}

\item{mm.2.go.df.MapManBin}{a list of the MapMan-Bins to analyze. Must have
the same length as \code{mm.2.go.df.MapManBin.GO}}
}
\value{
An instance of \code{data frame} with the following entries:
'mm.2.go.df.MapManBin' is the MapMan-Bin analyzed.
'n.words.shared' are the number of words shared between the MapMan-Bins
and their GOAs.
'percent.shared.words' is the percentage of shared words getween the
MapMan-Bin and its corresponding GOA.
'bins.gos.shared.words' is a list of words shared.
'bins.gos.no.shared.words' is a list of the words from the MapMan-Bin
description that are not shared with the GOA.
}
\description{
Analyze the number of shared words among the MapMan-Bins descriptions and the
GOs that conform their GOA.
}
