% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{annotateGeneFamily}
\alias{annotateGeneFamily}
\title{For the argument gene family find the most frequent Pfam domains and
mercator MapMan4 results, excluding argument BINCODES. Also report the
relative frequency, i.e. the fraction of genes sharing the respective
annotation.}
\usage{
annotateGeneFamily(gene.ids, fam.name, pfam.hmmer3.domtbl,
  mercator.map.man.4.tbl, mm4.excl.bincodes.regex = "^(35|50)")
}
\arguments{
\item{gene.ids}{- A character vector holding the gene identifiers of the
members of the argument family 'fam.name'.}

\item{fam.name}{- The name of the family in question.}

\item{pfam.hmmer3.domtbl}{- An instance of 'base::data.frame', the result of
invoking 'SlydGeneFamsAnalyses::parseHmmer3DomTableOut'.}

\item{mercator.map.man.4.tbl}{- An instance of 'base::data.frame', the
result of reading in Mercator MapMan4 annotatations for at least the genes
of the argument family. Read with 'mercator.map.man.4.tbl <- read.table(
'path.2.mercator.result.table.txt', header=T, sep='<TAB>', quote=''',
stringsAsFactors=F )'.}

\item{mm4.excl.bincodes.regex}{- A regular expression to be applied on the
'BINCODE' column of argument 'mercator.map.man.4.tbl' to exclude certain
MapMan4 BINS from consideration. Note that regular epression is expected to
be in PERL format. Default is '^(35|50)'.}
}
\value{
An instance of 'base::data.frame' with the following columns: Family
size anno.acc anno.desc rel.freq
}
\description{
For the argument gene family find the most frequent Pfam domains and
mercator MapMan4 results, excluding argument BINCODES. Also report the
relative frequency, i.e. the fraction of genes sharing the respective
annotation.
}
