% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManFunks.R
\name{appendToRData}
\alias{appendToRData}
\title{Enables appending to an existing RData file.}
\usage{
appendToRData(..., list = character(), file)
}
\arguments{
\item{...}{the R objects to append}

\item{list}{same as in \code{base::save}}

\item{file}{A valid file path to an RData file. If it does not exist, it
will be created only containing arguments \code{...}.}
}
\value{
The result of invoking \code{base::save}.
}
\description{
Enables appending to an existing RData file.
}
