% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManFunks.R
\name{compoundGoAnnotationEntropy}
\alias{compoundGoAnnotationEntropy}
\title{Infers the compound GO annotations found for the genes related to the
argument MapMan Bin and measures the resulting Shannon Entropy of these
compound GO annotations.}
\usage{
compoundGoAnnotationEntropy(map.man.bin,
  mm.bins.vs.genes = getOption("MapMan2GO.seq.sim.tbl",
  mm.bins.vs.sprot),
  mm.bin.col = getOption("MapMan2GO.seq.sim.tbl.bin.col", "MapManBin"),
  mm.gene.col = getOption("MapMan2GO.seq.sim.tbl.gene.col",
  "Swissprot.Short.ID"), goa.tbl = getOption("MapMan2GO.goa.tbl",
  ukb.goa.hits), gene.col = getOption("MapMan2GO.goa.tbl.gene.col", 3),
  go.col = getOption("MapMan2GO.goa.tbl.go.col", 2),
  extend.goas.with.ancestors = getOption("MapMan2GO.extend.goa.with.ancestors",
  TRUE), bp.gos = getOption("MapMan2GO.bp.gos", GO.BP),
  cc.gos = getOption("MapMan2GO.cc.gos", GO.CC),
  mf.gos = getOption("MapMan2GO.mf.gos", GO.MF))
}
\arguments{
\item{map.man.bin}{The identifier of the MapMan Bin to assign GO terms to.}

\item{mm.bins.vs.genes}{An instance of \code{data.frame} holding
MapManBin-Gene-Relations. Default is
\code{getOption('MapMan2GO.seq.sim.tbl', mm.bins.vs.sprot)}.}

\item{mm.bin.col}{The column of \code{mm.bins.vs.genes} in which to lookup
the MapMan-Bins. Default is \code{getOption('MapMan2GO.seq.sim.tbl.bin.col',
'MapManBin')}.}

\item{mm.gene.col}{The column of \code{mm.bins.vs.genes} in which to lookup
the gene identifiers. Default is
\code{getOption('MapMan2GO.seq.sim.tbl.gene.col', 'Swissprot.Short.ID')}.}

\item{goa.tbl}{An instance of \code{data.frame} holding GOAs. Default is
\code{getOption('MapMan2GO.goa.tbl', ukb.goa.hits)}}

\item{gene.col}{The column of \code{goa.tbl} number or name in which to
lookup the genes. Default is \code{getOption('MapMan2GO.goa.tbl.gene.col',
3)}.}

\item{go.col}{The column of \code{goa.tbl} number or name in which to lookup
the GO terms. Default is \code{getOption('MapMan2GO.goa.tbl.go.col',2)}}

\item{extend.goas.with.ancestors}{boolean indicating whether to extend each
proteins' GOA with the ancestors of the respective GO Terms. Default is
\code{getOption('MapMan2GO.extend.goa.with.ancestors', TRUE)}.}

\item{mm.bins.vs.genes}{An instance of \code{data.table} with at least two
columns. It must hold mappings of \code{map.man.bin} to genes at least
partially found in \code{goa.tbl}.}
}
\value{
An instance of \code{list} with the following named entries:
'Shannon.Entropy' is the measured entropy of compound GO annotations
retrieved for the genes related to \code{map.man.bin}. The second entry
'genes.goa' is a list with each genes' compound GO Annotations,
'MapManBin.GO' is the intersection of the genes' compound GO Annotations to
be used as the MapMan-Bin's compound GO Annotation, n.GO is the number of GO
Terms in the MapMan-Bin's compound GO Annotation, median.n.GO is the median
of the number of GO Terms in the genes' GOAs, and n.genes is the number of
genes related to \code{map.man.bin}.
}
\description{
Infers the compound GO annotations found for the genes related to the
argument MapMan Bin and measures the resulting Shannon Entropy of these
compound GO annotations.
}
