% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{createGeneFamilyWorkingDirIfNotExists}
\alias{createGeneFamilyWorkingDirIfNotExists}
\title{Creates the argument gene family 'gene.family.name' working directory in the
argument 'base.dir' if it does not already exists.}
\usage{
createGeneFamilyWorkingDirIfNotExists(base.dir, gene.family.name)
}
\arguments{
\item{base.dir}{- The valid file path to the base working directory}

\item{gene.family.name}{- The name of the gene family which defines the name
of the directory to be created within the argument 'base.dir'.}
}
\value{
The full path to the gene family's working directy, created if not
already existed.
}
\description{
Creates the argument gene family 'gene.family.name' working directory in the
argument 'base.dir' if it does not already exists.
}
