% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{findGenesWithPhysicoChemicalDivergentAA}
\alias{findGenesWithPhysicoChemicalDivergentAA}
\title{For a given MAPP result table find those argument genes of interest that
have a significantly divergent amino acid at the positions that MAPP
identified as divergent.}
\usage{
findGenesWithPhysicoChemicalDivergentAA(mapp.tbl, fam.aa.msa,
  genes.of.interest = names(slyd.cds), p.adjusted.cutoff = 0.05,
  mapp.aa.p.val.cols = setNames(c("A.1", "C.1", "D.1", "E.1", "F.1",
  "G.1", "H.1", "I.1", "K.1", "L.1", "M.1", "N.1", "P.1", "Q.1", "R.1",
  "S.1", "T.1", "V.1", "W.1", "Y.1"), c("A", "C", "D", "E", "F", "G", "H",
  "I", "K", "L", "M", "N", "P", "Q", "R", "S", "T", "V", "W", "Y")))
}
\arguments{
\item{mapp.tbl}{- The result of invoking
\code{SlydGeneFamsAnalyses::readMappResult}.}

\item{fam.aa.msa}{- The result of invoking
\code{SlydGeneFamsAnalyses::readMultipleSequenceAlignmentAsMatrix}.}

\item{genes.of.interest}{- A character vector of gene identifier. Set to
rownames(fam.aa.msa) if you want to get results for all genes in the
alignment. Default is \code{names(slyd.cds)}.}

\item{p.adjusted.cutoff}{- The cutoff for the adjusted P values to infer
significancy. Default is \code{.05}.}

\item{mapp.aa.p.val.cols}{- A named character or integer vector identifying
those columns of argument 'mapp.tbl' in which to lookup the respective amino
acid P values. Default is \code{setNames(c('A.1', 'C.1', 'D.1', 'E.1',
'F.1', 'G.1', 'H.1', 'I.1', 'K.1', 'L.1', 'M.1', 'N.1', 'P.1', 'Q.1', 'R.1',
'S.1', 'T.1', 'V.1', 'W.1', 'Y.1'), c('A', 'C', 'D', 'E', 'F', 'G', 'H',
'I', 'K', 'L', 'M', 'N', 'P', 'Q', 'R', 'S', 'T', 'V', 'W', 'Y'))}.}
}
\value{
An instance of \code{base::data.frame} with the following columns:
'Protein', 'Site', 'Divergent.AA', and 'AA.p.value'. Returns NULL if no
matches were found.
}
\description{
For a given MAPP result table find those argument genes of interest that
have a significantly divergent amino acid at the positions that MAPP
identified as divergent.
}
