% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{geneNameMappings}
\alias{geneNameMappings}
\title{Creates the argument gene family 'gene.family.name' mapping of original gene
identifiers to sanitized 'PROT1' ... 'PROTN' identifiers.}
\usage{
geneNameMappings(gene.family.dir, gene.family.name,
  gene.fams.lst = gene.families)
}
\arguments{
\item{gene.family.dir}{- The valid file path to the gene family's working
directory.}

\item{gene.family.name}{- The name of the gene family which defines the name
of the directory to be created within the argument 'base.dir'.}

\item{gene.fams.lst}{- The list of gene families in which to lookup the
argument 'gene.family.name'. Default is SlydGeneFamsAnalyses's
\code{gene.families}.}
}
\value{
A data.frame holding two columns: original and sanitized. The
dataframe is either loaded from a file 'name_mappings.txt' or created,
stored in that file, and then returned.
}
\description{
Creates the argument gene family 'gene.family.name' mapping of original gene
identifiers to sanitized 'PROT1' ... 'PROTN' identifiers.
}
