% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{generateInteractiveMsaPlot}
\alias{generateInteractiveMsaPlot}
\title{Generates an interactive plot of the family's multiple amino acid sequence
alignment. The output is an HTML file which requires this packages's
\code{./inst/assets} folder to be at the same location as the output file. A
modified version of the Javascript library
\code{github.com/AndrewCRMartin/JSAV} is used to display the alignment.
Warning: There are more beatifull ways to interactively show an alignment.}
\usage{
generateInteractiveMsaPlot(path.2.msa.fasta, meme.branches.tbl,
  pfam.hmmer3.domtbl, output.dir,
  brew.template = normalizePath(file.path(path.package("SlydGeneFamsAnalyses"),
  "msa_template_brew.html")))
}
\arguments{
\item{path.2.msa.fasta}{- The valid file path to the family's multiple amino
acid alignment in fasta format. The filename is also used to extract the
family's name and the name of the output file.}

\item{meme.branches.tbl}{- An instance of \code{base::data.frame} holding
the results of HyPhy's MEME analysis (branches). See
\code{SlydGeneFamsAnalyses::readMemeResults} for more details.}

\item{pfam.hmmer3.domtbl}{- An instance of \code{base::data.frame} holding
the results of running HMMER3 with the \code{--domtblout} command line
switch. Use \code{SlydGeneFamsAnalyses::parseHmmer3DomTableOut} to read in
the resulting table.}

\item{output.dir}{- The directory in which to write the resulting HTML file.}

\item{brew.template}{- The brew template to be used to render the HTML file.
Default is
\code{normalizePath(file.path(path.package('SlydGeneFamsAnalyses'),
'msa_template_brew.html'))}.}
}
\value{
The result of invoking \code{brew::brew(...)}.
}
\description{
Generates an interactive plot of the family's multiple amino acid sequence
alignment. The output is an HTML file which requires this packages's
\code{./inst/assets} folder to be at the same location as the output file. A
modified version of the Javascript library
\code{github.com/AndrewCRMartin/JSAV} is used to display the alignment.
Warning: There are more beatifull ways to interactively show an alignment.
}
