% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{generateMemeBatchFiles}
\alias{generateMemeBatchFiles}
\title{Generates the MEME HyPhy batch-files for all families found in
sub-directories of argument 'working.dir'.}
\usage{
generateMemeBatchFiles(working.dir, hyphy.batch.files.dir,
  gene.families.names = names(gene.families))
}
\arguments{
\item{working.dir}{- The valid file path to the directory in which to store
the sanitized and parsed MSAs. Each will be stored in its respective gene
family's sub-directory, identified by the family's name itself.}

\item{hyphy.batch.files.dir}{- The valid file path to the directory of the
HyPhy installation in which to find the HyPhy batch files.}

\item{gene.families.names}{- A character vector of gene family names that
are to be expected to contain the sub-directory names in argument
'working.dir' excluding the file-extensions. Default is this package's
\code{names(gene.families)}.}
}
\value{
TRUE if and only if no error has occurred.
}
\description{
Generates the MEME HyPhy batch-files for all families found in
sub-directories of argument 'working.dir'.
}
