% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManFunks.R
\name{goaForMapManBin}
\alias{goaForMapManBin}
\title{Uses the MapMan-Bin to Gene Ontology mappings generated within the
\code{MapMan2GO} package to assign a set of Gene Ontology terms to the
argument MapMan-Bin \code{mm.bincode}. Optional recursion looks up GO Terms
assigned to a parent MapMan-Bin, if no GO Terms are found for the argument
MapMan-Bin.}
\usage{
goaForMapManBin(mm.bincode,
  map.man.2.go = getOption("MapMan2GO.map.man.2.go", mm.2.go.df),
  use.parental.bins = getOption("MapMan2GO.use.parental.bins", FALSE),
  parental.regex = getOption("MapMan2GO.parental.regex", "\\\\.[0-9]+$"))
}
\arguments{
\item{mm.bincode}{A string identifying the MapMan-Bin, eg. \code{1.1.1.2}.}

\item{map.man.2.go}{An instance of \code{base::data.frame} holding the
MapMan-Bin to Gene Ontology mappings. The mapped GO Terms are expected to be
concatenated into a comma separated string. The table must have at least the
following two columns: \code{MapManBin} with the MapMan-Bin's identifiers
(BINCODE), and \code{MapManBin.GO} with the comma separated and concatenated
GO Terms assigned to the respective Bin. Default is
\code{getOption('MapMan2GO.map.man.2.go', mm.2.go.df)}.}

\item{use.parental.bins}{A BOOLEAN indicating wether to perform recursive
lookup of GO Terms assigned to parent Bins of \code{mm.bincode}, if
\code{mm.bincode} has no GO Terms assigned in \code{map.man.2.go}. Default
is \code{getOption('MapMan2GO.use.parental.bins', FALSE)}.}

\item{parental.regex}{A string representing the regular expresion to be used
to delete the matching trailing substring yielding the parental BINCODE of
\code{mm.bincode}. Default is \code{getOption('MapMan2GO.parental.regex',
'\\.[0-9]+$')}.}
}
\value{
A string of comma separated concatenated GO Terms assigned to the
argument \code{mm.bincode} or one of its ancestors (see argument
\code{use.parental.bins}). \code{NA} is returned, if no GO Terms can be
found.
}
\description{
Uses the MapMan-Bin to Gene Ontology mappings generated within the
\code{MapMan2GO} package to assign a set of Gene Ontology terms to the
argument MapMan-Bin \code{mm.bincode}. Optional recursion looks up GO Terms
assigned to a parent MapMan-Bin, if no GO Terms are found for the argument
MapMan-Bin.
}
