% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManFunks.R
\name{goCounts}
\alias{goCounts}
\title{Computes the empirical counts of events in a sample. Events are GO Term
annotations.}
\usage{
goCounts(go.sample.space, go.annos)
}
\arguments{
\item{go.sample.space}{A vector representing the sample space.}

\item{go.annos}{A vector of samples}
}
\value{
A named numeric of empirical counts, names are the events in
\code{go.sample.space} and values are the empirical counts as found in
\code{go.annos}.
}
\description{
Computes the empirical counts of events in a sample. Events are GO Term
annotations.
}
