% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManFunks.R
\name{goDepth}
\alias{goDepth}
\title{Counts the minimum number of edges from the Gene Ontology Graph's root node
to the argument node \code{x}.}
\usage{
goDepth(x, n = 1, go.ont = getOption("MapMan2GO.GO.Ontology", GO.OBO))
}
\arguments{
\item{x}{The node for which to compute its depth, e.g. 'GO:0044237'.}

\item{n}{The minimum number of edges any node has distance to the root node.
Default is \code{1}.}

\item{go.ont}{The result of \code{ontologyIndex::get_ontology('go.obo')}.
Default is \code{getOption('MapMan2GO.GO.Ontology', GO.OBO)}.}
}
\value{
An integer, the minimun number of edges from the GO Graph's root
node to the argument term \code{x}.
}
\description{
Counts the minimum number of edges from the Gene Ontology Graph's root node
to the argument node \code{x}.
}
