% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManFunks.R
\name{intersectAncestralEvidenceCodes}
\alias{intersectAncestralEvidenceCodes}
\title{Identifies the subset of evidence codes ancestral to \code{eco.id} that
intersect with argument \code{ancestral.eco.ids}.}
\usage{
intersectAncestralEvidenceCodes(eco.id,
  ontology = getOption("MapMan2GO.eco.ontology", ECO.OBO),
  ancestral.eco.ids = getOption("MapMan2GO.ancestral.eco.ids",
  eco.first.level))
}
\arguments{
\item{eco.id}{The evidence code's identifier for which to find intersecting
ancestors.}

\item{ontology}{An instance of class \code{ontologyIndex::ontology_index}
generated by \code{ontologyIndex::get_ontolgy('eco.obo'}. Default is
\code{getOption('MapMan2GO.eco.ontology', ECO.OBO)}.}

\item{ancestral.eco.ids}{A character vector of evidence code IDs to be
intersected with the ancestors of argument \code{eco.id}. Default is
\code{getOption('MapMan2GO.ancestral.eco.ids', eco.first.level)}.}
}
\value{
A character vector holding those terms ancestral to \code{eco.id}
found also in \code{ancestral.eco.ids}. \code{character(0)} is returned if
none are found.
}
\description{
Identifies the subset of evidence codes ancestral to \code{eco.id} that
intersect with argument \code{ancestral.eco.ids}.
}
