% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{loadAndSanitizeAAMsas}
\alias{loadAndSanitizeAAMsas}
\title{Loads all multiple amino acid sequence alignments (MSAs) for the respective
gene families reconstructed by Orthofinder. They are read in using
\code{seqinr::read.fasta} and gene identifiers are sanitized, removing the
species name prefixes added by Orthofinder. See function
\code{sanitizeOrthofinderGeneIds} for more details. Note, that this function
uses \code{mclapply}.}
\usage{
loadAndSanitizeAAMsas(path.2.msa.dir, working.dir,
  gene.families.names = names(gene.families))
}
\arguments{
\item{path.2.msa.dir}{- The valid file path to the directory the MSAs are
stored in. Make sure their file-extensions are '.fa'.}

\item{working.dir}{- The valid file path to the directory in which to store
the sanitized and parsed MSAs. Each will be stored in its respective gene
family's sub-directory, identified by the family's name itself.}

\item{gene.families.names}{- A character vector of gene family names that
are to be expected to contain the file names in argument 'path.2.msa.dir'
excluding the file-extensions. Default is this package's
\code{names(gene.families)}.}
}
\value{
TRUE if and only if no error has occurred.
}
\description{
Loads all multiple amino acid sequence alignments (MSAs) for the respective
gene families reconstructed by Orthofinder. They are read in using
\code{seqinr::read.fasta} and gene identifiers are sanitized, removing the
species name prefixes added by Orthofinder. See function
\code{sanitizeOrthofinderGeneIds} for more details. Note, that this function
uses \code{mclapply}.
}
