% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{mapOriginalToSanitizedNames}
\alias{mapOriginalToSanitizedNames}
\title{Returns the sanitized gene names mapped to the argument 'originals' as
assigned in argument 'mappings' table.}
\usage{
mapOriginalToSanitizedNames(originals, mappings)
}
\arguments{
\item{originals}{- Character vector of gene identifiers.}

\item{mappings}{- Instance of base::data.frame with two columns: 'original'
holding the original gene identifiers, and 'sanitized' holding the mapped
simplified gene identifiers. See function \code{geneNameMappings} for more
details.}
}
\value{
A character vector of same length as argument 'originals' holding
the respective sanitized gene identifiers.
}
\description{
Returns the sanitized gene names mapped to the argument 'originals' as
assigned in argument 'mappings' table.
}
