% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{mapSanitizedToOriginalNames}
\alias{mapSanitizedToOriginalNames}
\title{Returns the original gene names mapped to the argument 'sanitized' as
assigned in argument 'mappings' table.}
\usage{
mapSanitizedToOriginalNames(sanitized, mappings)
}
\arguments{
\item{sanitized}{- Character vector of gene identifiers.}

\item{mappings}{- Instance of base::data.frame with two columns: 'original'
holding the original gene identifiers, and 'sanitized' holding the mapped
simplified gene identifiers. See function \code{geneNameMappings} for more
details.}
}
\value{
A character vector of same length as argument 'sanitized' holding
the respective original gene identifiers.
}
\description{
Returns the original gene names mapped to the argument 'sanitized' as
assigned in argument 'mappings' table.
}
