% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{parseHmmer3DomTableOut}
\alias{parseHmmer3DomTableOut}
\title{Parses the output table of a HMMER3 run with command line switch
\code{--domtblout}.}
\usage{
parseHmmer3DomTableOut(path.2.hmmer3.domtbl.out)
}
\arguments{
\item{path.2.hmmer3.domtbl.out}{- Valid file path to the output of hmmer3
run with the command line switch \code{--domtblout}.}
}
\value{
An instance of \code{data.frame} with 23 columns containing the
output of HMMER3.
}
\description{
Parses the output table of a HMMER3 run with command line switch
\code{--domtblout}.
}
