% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{parseOrthogroupsTxt}
\alias{parseOrthogroupsTxt}
\title{Reads and parses the Orthofinder gene families output file.}
\usage{
parseOrthogroupsTxt(path.2.orthogroups.txt)
}
\arguments{
\item{path.2.orthogroups.txt}{- The valid file path to the Orthofinder
output 'orthogroups.txt'. Use the txt version, not the tsv, please.}
}
\value{
A list of gene families, names are the family names, and values are
the character vectors of each respective family's gene members.
}
\description{
Reads and parses the Orthofinder gene families output file.
}
