% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{printAaMsaWithSelection}
\alias{printAaMsaWithSelection}
\title{Uses Bioconductor's 'msa' package to pretty print the alignment of the
argument gene family highlighting the genes and amino acids found to be
subject to positive selection.}
\usage{
printAaMsaWithSelection(fam.dir, aa.msa.file, selected.pos.and.genes.tbl,
  plot.pdf.prefix = sub(".fa", ".pdf", aa.msa.file, fixed = TRUE),
  before.sel.pos.offset = 5, after.sel.pos.offset = 5)
}
\arguments{
\item{aa.msa.file}{- The name, not path, of the multiple sequence alignment
file to be used. It is of key importance that the protein identifiers in the
alignment file are identical to those in the argument
'selected.pos.and.genes.tbl' column 'Protein'.}

\item{selected.pos.and.genes.tbl}{- An instance of \code{base::data.frame}
holding information about positively selected proteins, aligned amino acid
positions and their corresponding unaligned positions. See function
\code{SlydGeneFamsAnalyses::readMemeResults} for details on how this table
can be generated.}

\item{plot.pdf.prefix}{- The file name with file extension ('.pdf') of the
plot that will be generated and saved in the argument 'fam.dir'. Default is
\code{sub('.fa', '.pdf', aa.msa.file, fixed=TRUE)}.}

\item{before.sel.pos.offset}{- The offset before the first selected position
to be included in the plot. Default is \code{5}.}

\item{after.sel.pos.offset}{- The offset after the first selected position
to be included in the plot. Default is \code{5}.}

\item{fam.fir}{- The path to the working directory of the argument gene
family. This dir should contain the multiple sequence alignment file and
will be used to write the alignment plots into.}
}
\value{
The result of invoking \code{msa::msaPrettyPrint(...)}.
}
\description{
Uses Bioconductor's 'msa' package to pretty print the alignment of the
argument gene family highlighting the genes and amino acids found to be
subject to positive selection.
}
