% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{readMappResult}
\alias{readMappResult}
\title{Simple wrapper function to read in MAPP result tables. See 'Stone and Sidow,
A. Physicochemical constraint violation by missense substitutions mediates
impairment of protein function and disease severity. Genome Res 15.}
\usage{
readMappResult(path.2.mapp.result.tbl, fam.name = NULL)
}
\arguments{
\item{path.2.mapp.result.tbl}{- The valid file path to the MAPP result
table.}

\item{fam.name}{- Optional argument string defining the name of the family
MAPP was executed for. If given the resulting data.frame will have an
additional column 'Family' holding this value. Default is 'NULL'.}
}
\value{
An instance of \code{base::data.frame} holding the contents of
argument 'path.2.mapp.result.tbl'.
}
\description{
Simple wrapper function to read in MAPP result tables. See 'Stone and Sidow,
A. Physicochemical constraint violation by missense substitutions mediates
impairment of protein function and disease severity. Genome Res 15.
}
