% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{readMemeResults}
\alias{readMemeResults}
\title{Within the argument 'work.dir' all gene family work dirs are looked up that
have HyPhy MEME result files. These are parsed and sites showing significant
evidence for positive selection are identified. Subsequently leaves of the
phylogenetic tree (proteins) that show significant evidence for positive
selection at the identified sites are also identified.}
\usage{
readMemeResults(work.dir, p.adjust.method = "fdr", pval.cutoff = 0.05,
  empirical.bayes.factor.cutoff = 100,
  phyl.tree.leaf.regex = "^PROT\\\\d+$",
  report.original.gene.names = TRUE,
  gene.family.sizes = gene.families.sizes)
}
\arguments{
\item{work.dir}{- The valid file path to the directory in which each
family's MEME results are stored.}

\item{p.adjust.method}{- The argument passed to \code{base::p.adjust} as the
method to apply for correcting p-values for multiple hypotheses testing.}

\item{pval.cutoff}{- The significance level for corrected p-values. Default
is \code{0.05}.}

\item{empirical.bayes.factor.cutoff}{- The significance level to be apliad
to empirical Bayes Factor values. Default is \code{100}.}

\item{phyl.tree.leaf.regex}{- The regular expression to be applied to
distinguish inner tree nodes from leaf nodes. Default is
\code{'^PROT\\d+$'}.}

\item{report.original.gene.names}{- Boolean switch to control whether the
genes found to be subject to positive selection whall be reported with their
original and not their sanitized names. See
\code{SolycGeneFams::mapOriginalToSanitizedNames} and
\code{SolycGeneFams::geneNameMappings} for more details. Default of this
argument is \code{TRUE}.}
}
\value{
An instance of \code{base::list} with names the families for which
MEME results were found and values lists with two entries 'MEME.sites' and
'MEME.branches', the respectively found subjects to positive selection
passing the significance levels.
}
\description{
Within the argument 'work.dir' all gene family work dirs are looked up that
have HyPhy MEME result files. These are parsed and sites showing significant
evidence for positive selection are identified. Subsequently leaves of the
phylogenetic tree (proteins) that show significant evidence for positive
selection at the identified sites are also identified.
}
