% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManFunks.R
\name{readMercatorResultTable}
\alias{readMercatorResultTable}
\title{Wraps \code{data.table::fread} to parse the result produced by 'Mercator'.}
\usage{
readMercatorResultTable(path.2.mercator.result.tbl,
  add.go.terms = getOption("MapMan2GO.add.GO.terms", TRUE),
  map.man.bins.2.go = getOption("MapMan2GO.map.man.bins.2.go",
  mm.2.go.df),
  sanitize.accession = getOption("MapMan2GO.read.mercator.sanitize.accession",
  FALSE))
}
\arguments{
\item{path.2.mercator.result.tbl}{The valid file path to the text file
holding the table produced as result by Mercator.}

\item{add.go.terms}{If set to \code{TRUE} the MapMan-Bin to Gene Ontology Term
mappings are looked up and automatically added as an additional column.
Default is \code{getOption('MapMan2GO.add.GO.terms', TRUE)}}

\item{map.man.bins.2.go}{An instance of \code{base::data.frame} holding the
MapMan-Bin to Gene Ontology Term mappings. It must have at least two
columns: One named 'MapManBin' holding the BINCODEs and another named
'MapManBin.GO' holding the compound GO Term annotations separated by comma.
Default is \code{getOption('MapMan2GO.map.man.bins.2.go', mm.2.go.df)}.}

\item{sanitize.accession}{boolean if TRUE the mercator table column
'IDENTIFIER' will be passed through
\code{toupper(MapMan2GO::sanitizeAccession)} and an additional column
'IDENTIFIER.LONG' will hold the long Uniprot protein identifiers. Default is
\code{getOption( 'MapMan2GO.read.mercator.sanitize.accession', FALSE)}.}
}
\value{
An instance of \code{data.table} holding the results of the mercator
annotation pipeline. Note, if you want the result to be a
\code{base::data.frame} call \code{as.data.frame} and the result.
}
\description{
Wraps \code{data.table::fread} to parse the result produced by 'Mercator'.
}
