% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{readMultipleSequenceAlignmentAsMatrix}
\alias{readMultipleSequenceAlignmentAsMatrix}
\title{Wrapper function that reads in a multiple sequence alignment and returns it
as a matrix. Leverages \code{seqinr::read.fasta} to read in the the MSA.}
\usage{
readMultipleSequenceAlignmentAsMatrix(path.2.msa, seqtype = "AA")
}
\arguments{
\item{path.2.msa}{- The valid file path to the respective MSA. Must be in
fasta format.}

\item{seqtype}{- The argument will directly passed to
\code{seqinr::reaad.fasta}. Default is \code{'AA'}.}
}
\value{
A character matrix representing the MSA. Rownames hold the gene
names.
}
\description{
Wrapper function that reads in a multiple sequence alignment and returns it
as a matrix. Leverages \code{seqinr::read.fasta} to read in the the MSA.
}
