% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManFunks.R
\name{sanitizeAccession}
\alias{sanitizeAccession}
\title{Extracts short UniProtKB gene accessions from the long version identifiers,
e.g. 'sp|Q8GWW7|AGUA_ARATH' will become 'Q8GWW7'.}
\usage{
sanitizeAccession(ukb.accs)
}
\arguments{
\item{ukb.accs}{A character vector of long UniProtKB gene accessions}
}
\value{
The short gene accessions extracted from the argument
\code{ukb.accs}.
}
\description{
Extracts short UniProtKB gene accessions from the long version identifiers,
e.g. 'sp|Q8GWW7|AGUA_ARATH' will become 'Q8GWW7'.
}
