% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{sanitizeOrthofinderGeneIds}
\alias{sanitizeOrthofinderGeneIds}
\title{Orthofinder puts the species name as a prefix before all gene identifiers.
This functions removes them by deleting everything and including to the last
'_' underscore character. Note, make sure your gene identifiers do NOT use
underscores.}
\usage{
sanitizeOrthofinderGeneIds(gene.ids)
}
\arguments{
\item{gene.ids}{- A character vector of gene identifiers modified by
Orthofinder.}
}
\value{
A modified version of the argument 'gene.ids'.
}
\description{
Orthofinder puts the species name as a prefix before all gene identifiers.
This functions removes them by deleting everything and including to the last
'_' underscore character. Note, make sure your gene identifiers do NOT use
underscores.
}
