% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManFunks.R
\name{shannonEntropy}
\alias{shannonEntropy}
\title{Computes the normalized empirical Shannon Entropy for counts delivered in
the argument \code{counts.table}. Basis of the \code{log} function is
natural and normalization is done by division by the maximum entropy
\code{log(length(counts.table))} (see
\href{https://en.wikipedia.org/wiki/Entropy <- (information <- theory)#Efficiency)}{Normalized
Entropy}.}
\usage{
shannonEntropy(counts.table,
  log.base = getOption("MapMan2GO.entropy.log.base", base::exp(1)))
}
\arguments{
\item{counts.table}{result of invoking \code{base::table(count.vector)}}

\item{log.base}{The base of the logarithm. Default is the natural logarithm,
\code{getOption('MapMan2GO.entropy.log.base', base::exp(1))}.}
}
\value{
A numeric value element [0,1], the normalized Shannon Entropy.
}
\description{
Computes the normalized empirical Shannon Entropy for counts delivered in
the argument \code{counts.table}. Basis of the \code{log} function is
natural and normalization is done by division by the maximum entropy
\code{log(length(counts.table))} (see
\href{https://en.wikipedia.org/wiki/Entropy <- (information <- theory)#Efficiency)}{Normalized
Entropy}.
}
