% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManFunks.R
\name{shannonEntropyForGoas}
\alias{shannonEntropyForGoas}
\title{Computes the Shannon Entropy for a set of compound Gene Ontology
Annotations, that is sets of GO Terms assigned to a number of genes. Each
distinct compound GOA is treated as a unique statistical event for which its
respective frequency is measured and used as basis for the entropy
calculation. Optionally each compound GOA can be intersected with an
arbitrary set of GO Terms, e.g. all molecular function (MF) terms, to
compute the entropy just within this sub-ontology of the Gene Ontology.}
\usage{
shannonEntropyForGoas(genes.goa,
  intersect.with = getOption("MapMan2GO.entropy.intersect.compound.GOA.with",
  NULL), entropy.funk = getOption("MapMan2GO.entropy.funk",
  MapMan2GO::shannonEntropy))
}
\arguments{
\item{genes.goa}{A list of character vectors, where each character vector
resembles a compound GO annotation. Names can be the corresponding gene
accessions.}

\item{intersect.with}{A character vector of an arbitrary set of Gene
Ontology Terms to intersect each gene's GOA with before computing the
entropy. Can be \code{GO.BP}, \code{GO.CC}, or \code{GO.MF} as examples. In
case of this argument having length zero no intersection will be done.
Default is \code{getOption('MapMan2GO.entropy.intersect.compound.GOA.with',
NULL)}.}

\item{entropy.funk}{The function to be used to compute the Shannon Entropy.
Can be e.g. \code{entropy::entropy} for the standard Shannon Entropy.
Default is the normalized Shannon Entropy
\code{getOption('MapMan2GO.entropy.funk', MapMan2GO::shannonEntropy)}.}
}
\value{
A numeric value the computed Shannon Entropy.
}
\description{
Computes the Shannon Entropy for a set of compound Gene Ontology
Annotations, that is sets of GO Terms assigned to a number of genes. Each
distinct compound GOA is treated as a unique statistical event for which its
respective frequency is measured and used as basis for the entropy
calculation. Optionally each compound GOA can be intersected with an
arbitrary set of GO Terms, e.g. all molecular function (MF) terms, to
compute the entropy just within this sub-ontology of the Gene Ontology.
}
