\name{SlydGeneFamsAnalyses-package}
\alias{SlydGeneFamsAnalyses-package}
\alias{SlydGeneFamsAnalyses}
\docType{package}
\title{
\packageTitle{SlydGeneFamsAnalyses}
}
\description{
\packageDescription{SlydGeneFamsAnalyses}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SlydGeneFamsAnalyses}
\packageIndices{SlydGeneFamsAnalyses}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{SlydGeneFamsAnalyses}

Maintainer: \packageMaintainer{SlydGeneFamsAnalyses}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
