% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManFunks.R
\name{splitMapManBinGOAs}
\alias{splitMapManBinGOAs}
\title{Splits the concatenated Gene Ontology Term Annotations (GOA) assigned to a
MapManBin into a vector.}
\usage{
splitMapManBinGOAs(map.man.bin.goas)
}
\arguments{
\item{map.man.bin.goas}{A characer vector of MapManBin-GOAs}
}
\value{
A character vector of sorted, unique GO Terms found in the comma
separated MapManBin-GOAs in argument \code{map.man.bin.goas}.
}
\description{
Splits the concatenated Gene Ontology Term Annotations (GOA) assigned to a
MapManBin into a vector.
}
