% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManFunks.R
\name{testGoaForMapManBin}
\alias{testGoaForMapManBin}
\title{Tests function \code{MapMan2GO::goaForMapManBin}.}
\usage{
testGoaForMapManBin()
}
\value{
\code{TRUE} if and only if all tests pass successfully.
}
\description{
Tests function \code{MapMan2GO::goaForMapManBin}.
}
