% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManFunks.R
\name{testGoDepth}
\alias{testGoDepth}
\title{Function to test \code{MapMan2GO::goDepth}.}
\usage{
testGoDepth()
}
\value{
\code{TRUE} if and only if all tests pass.
}
\description{
Function to test \code{MapMan2GO::goDepth}.
}
