% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapManFunks.R
\name{testShannonEntropy}
\alias{testShannonEntropy}
\title{Testing function \code{MapMan2GO::shannonEntropy}}
\usage{
testShannonEntropy()
}
\value{
\code{TRUE} if and only if all tests are passed successfully.
}
\description{
Testing function \code{MapMan2GO::shannonEntropy}
}
