% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{valdarMultipleAlignmentScore}
\alias{valdarMultipleAlignmentScore}
\title{Using William Valdar's method of scoring residues a multiple sequence
alignment is assigned a score ranging from 0 for low and 1 for high
conservation. The score is computed as the mean of each columns score, in
turn calculated as Score(column) = (1 -
normalizedShannonEntropy(column.non.gap.residues)) * (1 -
fraction.of.gaps(column)).
See \url{https://www.biostars.org/p/3856/} and
\url{https://doi.org/10.1002/prot.10146}.}
\usage{
valdarMultipleAlignmentScore(msa, gap.char = "-")
}
\arguments{
\item{msa}{- The multiple sequence alignment as a character matrix.}

\item{gap.char}{- The character representing gaps in the alignment. Default
is \code{'-'}.}
}
\value{
A numeric value the mean of the alignment's column scores.
}
\description{
Using William Valdar's method of scoring residues a multiple sequence
alignment is assigned a score ranging from 0 for low and 1 for high
conservation. The score is computed as the mean of each columns score, in
turn calculated as Score(column) = (1 -
normalizedShannonEntropy(column.non.gap.residues)) * (1 -
fraction.of.gaps(column)).
See \url{https://www.biostars.org/p/3856/} and
\url{https://doi.org/10.1002/prot.10146}.
}
